+++ material_3_demo/test/typography_screen_test.dart
@@ -1,16 +1,9 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:material_3_demo/main.dart';
 import 'package:material_3_demo/typography_screen.dart';
 
-import 'main_screen_test.dart';
+import 'component_screen_test.dart';
 
 void main() {
   testWidgets(
