+++ tool/flutter_ci_script_stable.sh
@@ -26,8 +26,8 @@ declare -ar PROJECT_NAMES=(
     "ios_app_clip"
     "isolate_example"
     "jsonexample"
-    # TODO: Re-enable once Dart 2.17 is stable
-    # "material_3_demo"
+    # TODO(domesticmouse): Re-enable once Dart 2.17 is stable
+    # "material_you_demo"
     "navigation_and_routing"
     "null_safety/null_safe_app"
     "null_safety/null_unsafe_app"
